fix: apply clang tidy to data provider files

+++ src/common/data_provider/src/hardware/architectureDependantX86_64Mac.cpp
@@ -12,5 +12,5 @@ int getMhz(IOsPrimitivesMac* osPrimitives)
         throw std::system_error {ret, std::system_category(), "Error reading cpu frequency."};
     }
 
-    return cpuHz / MHz;
+    return static_cast<int>(cpuHz / MHz);
 }
